merge code from filecoin-project

+++ cmd/lotus-bench/main.go
@@ -4,13 +4,16 @@ import (
 	"bytes"
 	"context"
 	"crypto/sha256"
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
+	"math/big"
 	"math/rand"
 	"os"
 	"path/filepath"
 	"time"
 
+	"github.com/docker/go-units"
 	ffi "github.com/filecoin-project/filecoin-ffi"
 	"github.com/ipfs/go-datastore"
 	logging "github.com/ipfs/go-log"
@@ -20,6 +23,8 @@ import (
 
 	"github.com/filecoin-project/lotus/build"
 	"github.com/filecoin-project/lotus/chain/address"
+	"github.com/filecoin-project/lotus/chain/types"
+	"github.com/filecoin-project/lotus/genesis"
 	"github.com/filecoin-project/lotus/lib/sectorbuilder"
 )
 
@@ -50,65 +55,101 @@ func main() {
 
 	log.Info("Starting lotus-bench")
 
+	build.SectorSizes = append(build.SectorSizes, 1024)
+
 	app := &cli.App{
 		Name:    "lotus-bench",
 		Usage:   "Benchmark performance of lotus on your hardware",
-		Version: build.Version,
+		Version: build.UserVersion,
 		Flags: []cli.Flag{
 			&cli.StringFlag{
 				Name:  "storage-dir",
 				Value: "~/.lotus-bench",
 				Usage: "Path to the storage directory that will store sectors long term",
 			},
-			&cli.Uint64Flag{
+			&cli.StringFlag{
 				Name:  "sector-size",
-				Value: 1024,
+				Value: "1GiB",
+				Usage: "size of the sectors in bytes, i.e. 32GiB",
 			},
 			&cli.BoolFlag{
 				Name:  "no-gpu",
 				Usage: "disable gpu usage for the benchmark run",
+			},
+			&cli.StringFlag{
+				Name:  "miner-addr",
+				Usage: "pass miner address (only necessary if using existing sectorbuilder)",
+				Value: "t0101",
+			},
+			&cli.StringFlag{
+				Name:  "benchmark-existing-sectorbuilder",
+				Usage: "pass a directory to run election-post timings on an existing sectorbuilder",
+			},
+			&cli.BoolFlag{
+				Name:  "json-out",
+				Usage: "output results in json format",
+			},
+			&cli.BoolFlag{
+				Name:  "skip-unseal",
+				Usage: "skip the unseal portion of the benchmark",
 			},
 		},
 		Action: func(c *cli.Context) error {
 			if c.Bool("no-gpu") {
 				os.Setenv("BELLMAN_NO_GPU", "1")
 			}
-			sdir, err := homedir.Expand(c.String("storage-dir"))
-			if err != nil {
-				return err
-			}
 
-			os.MkdirAll(sdir, 0775)
+			robench := c.String("benchmark-existing-sectorbuilder")
 
-			tsdir, err := ioutil.TempDir(sdir, "bench")
-			if err != nil {
-				return err
-			}
-			defer func() {
-				if err := os.RemoveAll(tsdir); err != nil {
-					log.Warn("remove all: ", err)
+			var sbdir string
+
+			if robench == "" {
+				sdir, err := homedir.Expand(c.String("storage-dir"))
+				if err != nil {
+					return err
 				}
-			}()
 
-			maddr, err := address.NewFromString("t0101")
+				os.MkdirAll(sdir, 0775)
+
+				tsdir, err := ioutil.TempDir(sdir, "bench")
+				if err != nil {
+					return err
+				}
+				defer func() {
+					if err := os.RemoveAll(tsdir); err != nil {
+						log.Warn("remove all: ", err)
+					}
+				}()
+				sbdir = tsdir
+			} else {
+				exp, err := homedir.Expand(robench)
+				if err != nil {
+					return err
+				}
+				sbdir = exp
+			}
+
+			maddr, err := address.NewFromString(c.String("miner-addr"))
 			if err != nil {
 				return err
 			}
 
-			sectorSize := c.Uint64("sector-size")
+			sectorSizeInt, err := units.RAMInBytes(c.String("sector-size"))
+			if err != nil {
+				return err
+			}
+			sectorSize := uint64(sectorSizeInt)
 
 			mds := datastore.NewMapDatastore()
 			cfg := &sectorbuilder.Config{
 				Miner:         maddr,
 				SectorSize:    sectorSize,
 				WorkerThreads: 2,
-				CacheDir:      filepath.Join(tsdir, "cache"),
-				SealedDir:     filepath.Join(tsdir, "sealed"),
-				StagedDir:     filepath.Join(tsdir, "staged"),
-				UnsealedDir:   filepath.Join(tsdir, "unsealed"),
+				Dir:           sbdir,
 			}
-			for _, d := range []string{cfg.CacheDir, cfg.SealedDir, cfg.StagedDir, cfg.UnsealedDir} {
-				if err := os.MkdirAll(d, 0775); err != nil {
+
+			if robench == "" {
+				if err := os.MkdirAll(sbdir, 0775); err != nil {
 					return err
 				}
 			}
@@ -126,7 +167,7 @@ func main() {
 			var sealTimings []SealingResult
 			var sealedSectors []ffi.PublicSectorInfo
 			numSectors := uint64(1)
-			for i := uint64(1); i <= numSectors; i++ {
+			for i := uint64(1); i <= numSectors && robench == ""; i++ {
 				start := time.Now()
 				log.Info("Writing piece into sector...")
 
@@ -181,17 +222,18 @@ func main() {
 
 				verifySeal := time.Now()
 
-				log.Info("Unsealing sector")
-				rc, err := sb.ReadPieceFromSealedSector(1, 0, dataSize, ticket.TicketBytes[:], commD[:])
-				if err != nil {
-					return err
-				}
+				if !c.Bool("skip-unseal") {
+					log.Info("Unsealing sector")
+					rc, err := sb.ReadPieceFromSealedSector(1, 0, dataSize, ticket.TicketBytes[:], commD[:])
+					if err != nil {
+						return err
+					}
 
+					if err := rc.Close(); err != nil {
+						return err
+					}
+				}
 				unseal := time.Now()
-
-				if err := rc.Close(); err != nil {
-					return err
-				}
 
 				sealTimings = append(sealTimings, SealingResult{
 					AddPiece:  addpiece.Sub(start),
@@ -206,6 +248,34 @@ func main() {
 
 			var challenge [32]byte
 			rand.Read(challenge[:])
+
+			if robench != "" {
+				// TODO: this assumes we only ever benchmark a preseal
+				// sectorbuilder directory... we need a better way to handle
+				// this in other cases
+
+				fdata, err := ioutil.ReadFile(filepath.Join(sbdir, "pre-seal-"+maddr.String()+".json"))
+				if err != nil {
+					return err
+				}
+
+				var genmm map[string]genesis.GenesisMiner
+				if err := json.Unmarshal(fdata, &genmm); err != nil {
+					return err
+				}
+
+				genm, ok := genmm[maddr.String()]
+				if !ok {
+					return xerrors.Errorf("preseal file didnt have expected miner in it")
+				}
+
+				for _, s := range genm.Sectors {
+					sealedSectors = append(sealedSectors, ffi.PublicSectorInfo{
+						CommR:    s.CommR,
+						SectorID: s.SectorID,
+					})
+				}
+			}
 
 			log.Info("generating election post candidates")
 			sinfos := sectorbuilder.NewSortedPublicSectorInfo(sealedSectors)
@@ -255,7 +325,7 @@ func main() {
 			}
 			verifypost2 := time.Now()
 
-			benchout := BenchResults{
+			bo := BenchResults{
 				SectorSize:     cfg.SectorSize,
 				SealingResults: sealTimings,
 
@@ -266,17 +336,28 @@ func main() {
 				VerifyEPostHot:         verifypost2.Sub(verifypost1),
 			} // TODO: optionally write this as json to a file
 
-			fmt.Println("results")
-			fmt.Printf("seal: addPiece: %s\n", benchout.SealingResults[0].AddPiece) // TODO: average across multiple sealings
-			fmt.Printf("seal: preCommit: %s\n", benchout.SealingResults[0].PreCommit)
-			fmt.Printf("seal: Commit: %s\n", benchout.SealingResults[0].Commit)
-			fmt.Printf("seal: Verify: %s\n", benchout.SealingResults[0].Verify)
-			fmt.Printf("unseal: %s\n", benchout.SealingResults[0].Unseal)
-			fmt.Printf("generate candidates: %s\n", benchout.PostGenerateCandidates)
-			fmt.Printf("compute epost proof (cold): %s\n", benchout.PostEProofCold)
-			fmt.Printf("compute epost proof (hot): %s\n", benchout.PostEProofHot)
-			fmt.Printf("verify epost proof (cold): %s\n", benchout.VerifyEPostCold)
-			fmt.Printf("verify epost proof (hot): %s\n", benchout.VerifyEPostHot)
+			if c.Bool("json-out") {
+				data, err := json.MarshalIndent(bo, "", "  ")
+				if err != nil {
+					return err
+				}
+
+				fmt.Println(string(data))
+			} else {
+				fmt.Printf("results (%d)\n", sectorSize)
+				if robench == "" {
+					fmt.Printf("seal: addPiece: %s (%s)\n", bo.SealingResults[0].AddPiece, bps(bo.SectorSize, bo.SealingResults[0].AddPiece)) // TODO: average across multiple sealings
+					fmt.Printf("seal: preCommit: %s (%s)\n", bo.SealingResults[0].PreCommit, bps(bo.SectorSize, bo.SealingResults[0].PreCommit))
+					fmt.Printf("seal: commit: %s (%s)\n", bo.SealingResults[0].Commit, bps(bo.SectorSize, bo.SealingResults[0].Commit))
+					fmt.Printf("seal: verify: %s\n", bo.SealingResults[0].Verify)
+					fmt.Printf("unseal: %s  (%s)\n", bo.SealingResults[0].Unseal, bps(bo.SectorSize, bo.SealingResults[0].Unseal))
+				}
+				fmt.Printf("generate candidates: %s (%s)\n", bo.PostGenerateCandidates, bps(bo.SectorSize*uint64(len(bo.SealingResults)), bo.PostGenerateCandidates))
+				fmt.Printf("compute epost proof (cold): %s\n", bo.PostEProofCold)
+				fmt.Printf("compute epost proof (hot): %s\n", bo.PostEProofHot)
+				fmt.Printf("verify epost proof (cold): %s\n", bo.VerifyEPostCold)
+				fmt.Printf("verify epost proof (hot): %s\n", bo.VerifyEPostHot)
+			}
 			return nil
 		},
 	}
@@ -285,4 +366,11 @@ func main() {
 		log.Warn(err)
 		return
 	}
+}
+
+func bps(data uint64, d time.Duration) string {
+	bdata := new(big.Int).SetUint64(data)
+	bdata = bdata.Mul(bdata, big.NewInt(time.Second.Nanoseconds()))
+	bps := bdata.Div(bdata, big.NewInt(d.Nanoseconds()))
+	return (types.BigInt{bps}).SizeStr() + "/s"
 }
